+++ inc/core/poppackethandler.hpp
@@ -20,7 +20,7 @@ class PopArtemisRPC;
 class PopS3pRPC;
 class PopChannelMap;
 
-class PopPacketHandler : public PopSink<uint64_t>
+class PopPacketHandler : public PopSink<uint32_t>
 {
 public:
 	struct PopPacketQueue
@@ -32,7 +32,7 @@ public:
 
 
 	PopPacketHandler(unsigned notused);
-	void process(const uint64_t* data, size_t size, const PopTimestamp* timestamp_data, size_t timestamp_size);
+	void process(const uint32_t* data, size_t size, const PopTimestamp* timestamp_data, size_t timestamp_size);
 	void process_ota_packet(ota_packet_t* p, uint32_t txTime, uint64_t pitTxTime, uint64_t pitPrnCodeStart);
 	void execute(const struct json_token *methodTok, const struct json_token *paramsTok, const struct json_token *idTok, struct json_token arr[POP_JSON_RPC_SUPPORTED_TOKENS], char *str, uint32_t txTime, uint64_t pitTxTime, uint64_t pitPrnCodeStart);
 	int32_t pop_get_tracker_slot_now(uuid_t uuid);
